# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "École Platform API",
    "description": "EdTech SaaS platform API for K-12 schools in Morocco.\n\n## Domains\n- **IAM** — Authentication, authorization, invitations, recovery, 2FA\n- **ERP** — Classes, enrollments, teacher assignments, attendance\n- **LMS** — Courses, assignments, submissions, results, content, activities, assessments\n- **Billing** — Invoices and payments\n- **COM** — Notifications, consents, parent feed\n- **AI** — Writing assistance, recommendations, KPIs\n\n## Authentication\nAll protected endpoints require a `Bearer` access token in the `Authorization` header.\nTokens are obtained via `POST /auth/login`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/recovery/request": {
      "post": {
        "tags": [
          "recovery"
        ],
        "summary": "Request password recovery",
        "description": "Request account recovery. Always returns 200 to prevent email enumeration.\n\nIf user exists: creates recovery request and generates OTP (logged in dev).\nIf user doesn't exist: returns success anyway.",
        "operationId": "request_recovery_api_v1_recovery_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Recovery request ID (always 200, no email enumeration)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recovery/verify": {
      "post": {
        "tags": [
          "recovery"
        ],
        "summary": "Verify recovery OTP",
        "description": "Verify recovery OTP. Transitions status: pending -> verified.\n\nLockout after 5 failed attempts (30-minute lock).",
        "operationId": "verify_recovery_api_v1_recovery_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP verification confirmation",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recovery/reset": {
      "post": {
        "tags": [
          "recovery"
        ],
        "summary": "Reset password",
        "description": "Reset user's password. Requires status=verified.\n\nRevokes all active sessions for the user (force re-login).\nState machine: pending -> verified -> reset (no backward transitions).",
        "operationId": "reset_password_api_v1_recovery_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoveryResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password reset confirmation",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments/initiate": {
      "post": {
        "tags": [
          "billing-payments"
        ],
        "summary": "Initiate a payment",
        "description": "Initiate a payment attempt for an invoice.\n\nValidates:\n1. Invoice exists, is in same school, belongs to the parent\n2. Invoice is in pending status\n3. Idempotency via idempotency_key (returns existing if same key)",
        "operationId": "initiate_payment_api_v1_payments_initiate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentInitiateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Payment attempt with provider redirect",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/payments/{attempt_id}": {
      "get": {
        "tags": [
          "billing-payments"
        ],
        "summary": "Get payment status",
        "description": "Get payment attempt status.",
        "operationId": "get_payment_status_api_v1_payments__attempt_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment attempt status",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/webhook/provider": {
      "post": {
        "tags": [
          "billing-payments"
        ],
        "summary": "Handle provider webhook",
        "description": "Handle a webhook event from a payment provider.\n\nINV-BIL-WEBHOOK: duplicate provider_event_id = no-op (idempotent).\n\nFlow:\n1. Check for duplicate provider_event_id\n2. Find matching payment attempt (optional)\n3. Update payment attempt status if found\n4. Update invoice status if payment is finalized\n5. Store webhook event",
        "operationId": "handle_provider_webhook_api_v1_payments_webhook_provider_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Webhook processing result",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentInitiateRequest": {
        "properties": {
          "invoice_id": {
            "type": "string",
            "format": "uuid",
            "title": "Invoice Id"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "invoice_id",
          "idempotency_key"
        ],
        "title": "PaymentInitiateRequest"
      },
      "RecoveryRequestCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "school_id": {
            "type": "string",
            "format": "uuid",
            "title": "School Id"
          }
        },
        "type": "object",
        "required": [
          "email",
          "school_id"
        ],
        "title": "RecoveryRequestCreate"
      },
      "RecoveryResetRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "new_password": {
            "type": "string",
            "minLength": 12,
            "title": "New Password",
            "description": "New password (min 12 chars, Phase 2A policy)"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "new_password"
        ],
        "title": "RecoveryResetRequest"
      },
      "RecoveryVerifyRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "otp": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "title": "Otp",
            "description": "6-digit OTP"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "otp"
        ],
        "title": "RecoveryVerifyRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebhookEventRequest": {
        "properties": {
          "provider_event_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Provider Event Id"
          },
          "payment_attempt_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Attempt Id"
          },
          "event_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Event Type"
          },
          "status": {
            "type": "string",
            "pattern": "^(paid|failed|canceled)$",
            "title": "Status"
          },
          "signature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature"
          }
        },
        "type": "object",
        "required": [
          "provider_event_id",
          "event_type",
          "status"
        ],
        "title": "WebhookEventRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}